fix(t145): move PowerShell BOM gate into electron-free shared module
CI desktop contracts failed importing package-portable.mjs because it pulls
@electron/asar from desktop/electron/node_modules (not installed on runners).
assertPowerShellBom now lives in desktop/shared/src/powershell-bom.ts with no
electron dependency; package-portable.mjs imports it. Packaged Depot bytes
unchanged (cf7f2b2c).
